fix(temporal): break namespace migration audit cycle
Why: The namespace migration audit and migration implementation imported each other eagerly, causing the required architecture check to fail on PR #2525.

What: Move the audit input and migration type definitions into a dependency-neutral module while preserving the migration module's public type API.

Verification: bunx turbo run typecheck test lint --filter=@shepherdjerred/temporal --filter=@homelab/cdk8s
